Moved all v1 code to separate dir and revert deprecations
