+++ packages/block-library/src/gallery/save.js
@@ -7,7 +7,7 @@ import { RichText, useBlockProps, InnerBlocks } from '@wordpress/block-editor';
  * Internal dependencies
  */
 import { defaultColumnsNumber } from './shared';
-import saveV1 from './save-v1';
+import saveV1 from './v1/save';
 
 export default function save( { attributes } ) {
 	if ( attributes?.ids?.length > 0 ) {
