+++ packages/block-library/src/gallery/v1/edit.js
@@ -40,15 +40,15 @@ import { View } from '@wordpress/primitives';
 /**
  * Internal dependencies
  */
-import { sharedIcon } from './shared-icon';
-import { pickRelevantMediaFiles } from './shared';
-import { defaultColumnsNumberV1 } from './deprecated';
-import Gallery from './gallery-v1';
+import { sharedIcon } from '../shared-icon';
+import { pickRelevantMediaFiles } from '../shared';
+import { defaultColumnsNumberV1 } from '../deprecated';
+import Gallery from './gallery';
 import {
 	LINK_DESTINATION_ATTACHMENT,
 	LINK_DESTINATION_MEDIA,
 	LINK_DESTINATION_NONE,
-} from './constants';
+} from '../constants';
 
 const MAX_COLUMNS = 8;
 const linkOptions = [
